Title: 컴퓨터와 함께하는 숫자 게임 2, Time: 135ms, Memory: 9MB, Status: Passed - CodeTree

+++ 241022/컴퓨터와 함께하는 숫자 게임 2/play-number-game-with-computer-2.java
@@ -40,10 +40,10 @@ public class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st = new StringTokenizer(br.readLine());
 
-        m = Integer.parseInt(st.nextToken());
+        m = Long.parseLong(st.nextToken());
         
         st = new StringTokenizer(br.readLine());
-        s = Integer.parseInt(st.nextToken());
-        e = Integer.parseInt(st.nextToken());
+        s = Long.parseLong(st.nextToken());
+        e = Long.parseLong(st.nextToken());
     }
 }
